+++ packages/managers/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @module-federation/managers
 
+## 2.6.0
+
+### Patch Changes
+
+- @module-federation/sdk@2.6.0
+
 ## 2.5.1
 
 ### Patch Changes